weather_api/requirement.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
Django==5.1.5
django-filter==25.1
django-taggit==6.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
mypy-extensions==1.0.0
mysql-connector==2.2.9
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg==3.2.6
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0